refactor AddResourceButton to HeaderButton
